feat(ios): PDF reader view + factory dispatch on publication type
Phase 2 of the PDF support roadmap. Splits the iOS reader into an
EPUB and a PDF implementation, switched by the factory based on the
current publication.

- Extract a `ReadiumReaderView` Swift protocol covering the surface
  that `FlutterReadiumPlugin` already invokes on the active reader
  view (getCurrentLocation, getFirstVisibleLocator, goToLocator,
  goToProgression, syncToLocator, applyDecorations). The plugin now
  holds `(any ReadiumReaderView)?` rather than a concrete class.
- Rename the existing `ReadiumReaderView` class to `EPUBReaderView`
  (file renamed; class adopts the protocol — no behavioural change).
  Status-emission helpers and the reader bug logger move to the new
  `ReadiumReaderView.swift` so both classes share them.
- Add `PDFReaderView` wrapping `PDFNavigatorViewController`. It wires
  the standard navigator delegate callbacks (locationDidChange,
  didJumpTo, presentError, didFailToLoadResourceAt, presentExternalURL)
  and the Flutter method channel (go / goForward / goBackward / dispose).
  `applyDecorations` and `syncToLocator` are intentional no-ops —
  `PDFNavigatorViewController` does not conform to `DecorableNavigator`
  in swift-toolkit 3.7.0 and PDFs have no media-overlay sync.
  `setPreferences` is a no-op stub pending Phase 5.
- `ReadiumReaderViewFactory.create` now inspects the current
  publication's `Publication.Profile.pdf` conformance (with the first
  reading-order entry's `MediaType.pdf` as a fallback) and returns
  `PDFReaderView` or `EPUBReaderView` accordingly. The Flutter platform
  view identifier is unchanged, so Dart sees a single widget type.

PDF rendering on iOS now reaches the screen for `application/pdf`
publications; Android navigator wiring follows in Phase 3.
